requirements.txt
amazon-braket-sdk>=1.0.0